+++ .example.env
@@ -1,0 +1,2 @@
+VITE_BACKEND_HOST=http://127.0.0.1
+VITE_BACKEND_PORT=8080
