+++ .github/workflows/release.yml
@@ -1,22 +1,55 @@
-name: Release
+name: Build and Release
 
 on:
   push:
-    tags:
-      - "v*"
+    branches: [main]
+  pull_request:
+    branches: [main]
+    types:
+      - closed
 
 jobs:
-  release:
+  image-release:
+    if: github.event.pull_request.merged == true || github.event_name == 'push'
     runs-on: ubuntu-latest
+
     steps:
-      - uses: actions/checkout@v3
+      - name: Login to GitHub Container Registry
+        uses: docker/login-action@v1
         with:
-          fetch-depth: 0
+          registry: ghcr.io
+          username: name-ethnicity-classifier
+          password: ${{ secrets.GITHUB_TOKEN }}
 
-      - uses: actions/setup-node@v3
-        with:
-          node-version: 16.x
+      - name: Checkout repository
+        uses: actions/checkout@v2
 
-      - run: npx changelogithub
+      - name: Generate version based on date and timestamp
+        id: version
+        run: |
+          VERSION=$(date +'%d-%m-%Y-%s')
+          echo "VERSION=${VERSION}" >> $GITHUB_ENV
+
+      - name: Push image to GitHub Container Registry
         env:
-          GITHUB_TOKEN: ${{secrets.GITHUB_TOKEN}}
+            BACKEND_HOST: "https://api.name-to-ethnicity.com"
+            BACKEND_PORT: "8080"
+        run: |
+          IMAGE_NAME=ghcr.io/name-ethnicity-classifier/n2e-frontend
+          docker build \
+            --build-arg VITE_BACKEND_HOST=${{ env.BACKEND_HOST }} \
+            --build-arg VITE_BACKEND_PORT=${{ env.BACKEND_PORT }} \
+            -t $IMAGE_NAME:$VERSION -t $IMAGE_NAME:latest .
+          
+          docker push $IMAGE_NAME:$VERSION
+          docker push $IMAGE_NAME:latest
+
+      - name: Create Release
+        uses: actions/create-release@v1
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          tag_name: ${{ env.VERSION }}
+          release_name: 🎉 N2E Frontend ${{ env.VERSION }}
+          draft: false
+          prerelease: false
