+++ .gitignore
@@ -22,3 +22,5 @@ coverage/
 
 # ESLint
 .eslintcache
+
+.env
