+++ src/lib/pages/model-hub/components/ModelSelectionList.tsx
@@ -5,6 +5,7 @@ import { useAuth } from "~/lib/contexts/AuthContext";
 import RequestModelModal from "./RequestModelModal";
 import { LuLock } from "react-icons/lu";
 import Cookies from "js-cookie";
+import { acessAlertToast } from "~/lib/utils/toasts";
 
 
 const MODEL_COLORS = ["#FFA7A7", "#F8D78F", "#8AB0F5", "#91E489", "#CCA4EF", "#F396B0", "#85DCEC", "#EBEB75", "#ED94EB", "#AA8EEA"];
@@ -39,7 +40,7 @@ const ModelSelectionList = (props: ModelSelectionProps) => {
                             width="full"
                             margin="auto"
                             leftIcon={<LuLock />}
-                            onClick={() => {}}
+                            onClick={() => acessAlertToast(toast)}
                         >
                             Access not yet granted.
                         </Button>
